RsID time out
  * Changed secondary time out to use real time vice
    RsID symbols
    - required when actual modem sample rate is greater
      than RsID sample rate, such as 8PSK modem types

+++ src/include/rsid.h
@@ -96,7 +96,7 @@ private:
 	static const RSIDs  rsid_ids_2[];
 	static const int rsid_ids_size2;
 
-	int rsid_secondary_time_out;
+	double rsid_secondary_time_out;
 
 	int hamming_resolution;
